Clean up yarpc package
